bugfix: truncating all tables probably shouldn't truncate the special 'schema_migrations' table
git-svn-id: file:///data/subversion/eol/trunk@219 78829999-583a-0410-bd01-8a9b849fd409

+++ spec/eol_spec_helpers.rb
@@ -13,8 +13,10 @@ module EOL::Spec
       options[:verbose] ||= false
       all_connections.each do |conn|
         conn.tables.each   do |table|
-          puts "[#{conn.instance_eval { @config[:database] }}].`#{table}`" if options[:verbose]
-          conn.execute "TRUNCATE TABLE`#{table}`"
+          unless table == 'schema_migrations'
+            puts "[#{conn.instance_eval { @config[:database] }}].`#{table}`" if options[:verbose]
+            conn.execute "TRUNCATE TABLE`#{table}`"
+          end
         end
       end
     end
